pyproject.toml
[build-system]
requires = ["flit_core >=3.4.0,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "geb-model"
dist-name = "geb-model"
authors = [
    {name = "Jens de Bruijn", email = "jens.de.bruijn@vu.nl"},
    {name = "Maurice Kalthof", email = "w.m.l.kalthof@vu.nl"},
    {name = "Veerle Bril", email = "v.c.bril@vu.nl"},
    {name = "Tim Busker", email = "tim.busker@vu.nl"},
]
dependencies = [
    "cdsapi>=0.7.0",
    "dask>=2023.3.0",
    "deap>=1.4.1",
    "flopy>=3.8.0",
    "geopandas>=1.0.0",
    "honeybees>=1.0.3",
    "hydromt==0.10.0",
    "isimip-client>=1.0.1",
    "matplotlib>=3.9.2",
    "numba>=0.60.0",
    "numpy>=2.0.1",
    "openpyxl>=3.1.2",
    "pybind11>=2.13.4",
    "pyogrio>=0.9.0",
    "pyyaml>=6.0.2",
    "rasterio>=1.3.10",
    "rioxarray>=0.15",
    "s3fs>=2023.6.0",
    "SALib>=1.4.8",
    "scipy>=1.14.0",
    "seaborn>=0.13.2",
    "tbb4py>=2021.13.1",
    "tqdm>=4.66.1",
    "xarray>=2024.3.0",
    "xclim>=0.52.0",
    "xesmf>=0.8.4",
    "xmipy>=1.5.0",
    "zarr>=2.18.2"
]
requires-python = ">=3.11, <3.13"
readme = "README.md"
classifiers = [
    # https://pypi.python.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Hydrology",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
]
dynamic = ['version', 'description']

[tool.flit.module]
name = "geb"

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]
docs = [
    "sphinx<8",
    "sphinx-autodoc-typehints",
    "sphinxcontrib-autoprogram",
    "sphinx_rtd_theme",
    "sphinxcontrib-autoyaml"
]
gpu = [
    "cupy"
]

full = [
]

publish = [
    "build",
    "twine",
]

fairstream = [
    "pgmpy>=0.1.24",
    "daft>=0.1.2",
    "opt-einsum>=3.3.0",
]

[project.scripts]
"geb" = "geb.cli:main"